+++ platform/jvm/capture/src/main/kotlin/io/bitdrift/capture/network/okhttp/CaptureOkHttpTracingInterceptor.kt
@@ -10,7 +10,6 @@ package io.bitdrift.capture.network.okhttp
 import androidx.annotation.VisibleForTesting
 import io.bitdrift.capture.Capture
 import io.bitdrift.capture.CaptureRuntimeProvider
-import io.bitdrift.capture.IInternalLogger
 import io.bitdrift.capture.ILogger
 import io.bitdrift.capture.IRuntimeProvider
 import io.bitdrift.capture.LogLevel
@@ -105,7 +104,7 @@ class CaptureOkHttpTracingInterceptor
         }
 
         private fun TracePropagationMode.logInternalStatus() {
-            val internalLogger = Capture.logger() as? IInternalLogger
+            val internalLogger = Capture.getInternalLogger()
             internalLogger?.logInternal(
                 type = LogType.INTERNALSDK,
                 level = LogLevel.INFO,