+++ platform/jvm/gradle-test-app/src/main/java/io/bitdrift/gradletestapp/ui/compose/components/TestingToolsCard.kt
@@ -30,6 +30,8 @@ fun TestingToolsCard(
     onLogSingleMessage: () -> Unit,
     onLogManyMessages: () -> Unit,
     onLogJsonField: () -> Unit,
+    onStartSpan: () -> Unit,
+    onEndSpan: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
     BdSectionCard(
@@ -59,6 +61,18 @@ fun TestingToolsCard(
         BdSecondaryButton(
             text = "Log JSON Field",
             onClick = onLogJsonField,
+            modifier = Modifier.fillMaxWidth(),
+        )
+
+        BdPrimaryButton(
+            text = "Start Span",
+            onClick = onStartSpan,
+            modifier = Modifier.fillMaxWidth(),
+        )
+
+        BdSecondaryButton(
+            text = "End Span",
+            onClick = onEndSpan,
             modifier = Modifier.fillMaxWidth(),
         )
     }